refactor: change name of explore genre card and grid to Popular genre card and grid
