+++ screens/HomeScreen.tsx
@@ -4,7 +4,8 @@ import ConcertList from "../components/layouts/ConcertList";
 import BottomNavBar from "../components/nav-bar/BottomNavBar";
 import { useUserStore } from "../stores/useUserStore";
 import { useLocationStore } from "../stores/useLocationStore";
-import ExploreGenresGrid from "../components/layouts/ExploreGenresGrid";
+
+import PopularGenresGrid from "../components/layouts/PopularGenresGrid";
 
 export default function HomeScreen() {
   const { userData } = useUserStore();
@@ -23,7 +24,7 @@ export default function HomeScreen() {
         <Text style={styles.sectionTitle}>Concerts in {city}</Text>
         <ConcertList />
         <Text style={styles.sectionTitle}>Popular Genres</Text>
-        <ExploreGenresGrid />
+        <PopularGenresGrid />
       </SafeAreaView>
       <BottomNavBar />
     </View>
